bugfix after launching monkey on my code

+++ convertion/ft_cmin_conv.c
@@ -6,7 +6,7 @@
 /*   By: mrozniec <mrozniec@student.le-101.fr>      +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/11/29 08:52:17 by mrozniec     #+#   ##    ##    #+#       */
-/*   Updated: 2019/12/13 00:35:59 by mrozniec    ###    #+. /#+    ###.fr     */
+/*   Updated: 2019/12/13 07:30:13 by mrozniec    ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -16,7 +16,7 @@
 char	*ft_cmin_conv(t_printf *wip)
 {
 	char	*res;
-	size_t	size_data;
+	int	size_data;
 
 	if ((wip->flags & L_MIN) != 0)
 		res = ft_lflag(wip);
